# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "u-cafe",
        "description": "Documentation for u-cafe admin dashboard and main site",
        "contact": {
            "email": "s.murodjonov@uzinfocom.uz"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/menu-type": {
            "get": {
                "tags": [
                    "MenuTypes"
                ],
                "summary": "Get list of MenuType",
                "description": "Returns list of MenuType",
                "operationId": "indexMenuType",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MenuTypes"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "MenuTypes"
                ],
                "summary": "Insert a new MenuType",
                "description": "Insert a new MenuType",
                "operationId": "storeMenuType",
                "requestBody": {
                    "description": "MenuType to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/MenuType"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "MenuType created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MenuType"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/menu-type/{menutype_id}": {
            "get": {
                "tags": [
                    "MenuTypes"
                ],
                "summary": "Show a MenuType from his Id",
                "description": "Show a MenuType from his Id",
                "operationId": "showMenuType",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/MenuType--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MenuType"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "MenuType not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "MenuTypes"
                ],
                "summary": "Delete a MenuType",
                "description": "Delete a MenuType",
                "operationId": "destroyMenuType",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/MenuType--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MenuType"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "MenuType not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "MenuTypes"
                ],
                "summary": "Update an existing MenuType",
                "description": "Update an existing MenuType",
                "operationId": "updateMenuType",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/MenuType--id"
                    }
                ],
                "requestBody": {
                    "description": "MenuType to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/MenuType"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MenuType"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "MenuType not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MenuType": {
                "title": "MenuType",
                "description": "MenuType model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of MenuType",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of MenuType",
                        "type": "string",
                        "example": "Macbook Pro"
                    },
                    "sku": {
                        "title": "sku",
                        "description": "sku of MenuType",
                        "type": "string",
                        "example": "MCBPRO2022"
                    },
                    "price": {
                        "title": "price",
                        "description": "price of MenuType",
                        "type": "integer",
                        "example": "99"
                    },
                    "created_at": {
                        "title": "created_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    }
                },
                "type": "object"
            },
            "MenuTypes": {
                "title": "MenuTypes list",
                "properties": {
                    "data": {
                        "title": "data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MenuType"
                        }
                    },
                    "first_page_url": {
                        "title": "first_page_url",
                        "type": "string",
                        "example": "http://localhost:8080/api/merchant-customers?page=1"
                    },
                    "last_page_url": {
                        "title": "last_page_url",
                        "type": "string",
                        "example": "http://localhost:8080/api/merchant-customers?page=3"
                    },
                    "prev_page_url": {
                        "title": "prev_page_url",
                        "type": "string",
                        "example": "null"
                    },
                    "next_page_url": {
                        "title": "next_page_url",
                        "type": "string",
                        "example": "http://localhost:8080/api/merchant-customers?page=2"
                    },
                    "current_page": {
                        "title": "current_page",
                        "type": "integer",
                        "example": "1"
                    },
                    "from": {
                        "title": "from",
                        "type": "integer",
                        "example": "1"
                    },
                    "last_page": {
                        "title": "last_page",
                        "type": "integer",
                        "example": "3"
                    },
                    "path": {
                        "title": "path",
                        "type": "string",
                        "example": "http://localhost:8080/api/merchant-customers"
                    },
                    "per_page": {
                        "title": "per_page",
                        "type": "integer",
                        "example": "1"
                    },
                    "total": {
                        "title": "total",
                        "type": "integer",
                        "example": "3"
                    },
                    "links": {
                        "title": "links",
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "title": "url",
                                    "type": "string",
                                    "example": "http://localhost:8080/api/merchant-customers?page=2"
                                },
                                "label": {
                                    "title": "label",
                                    "type": "string",
                                    "example": "1"
                                },
                                "active": {
                                    "title": "active",
                                    "type": "boolean",
                                    "example": "true"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
...
            "MenuType--id": {
                "name": "MenuType_id",
                "in": "path",
                "description": "Id of MenuType",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "1"
                }
            },
...
        }
    }
}